feat: move managed writer test to system test
