+++ system-test/fixtures/managedwriter/test_protos/customer_record.d.ts
