+++ system-test/fixtures/managedwriter/test_protos/index.ts
